Swap event and attributes
Clients remain the same, but the payload is json-encoded and send as
the body of the message, whereas the "event" is registered in an
attribute.

This is done because of two reasons:

- attributes are limited to 100, and that may fall sort for some use
cases

- more importantly, attributes are stored and dealt with as plain
strings, and may pose some challenges de/serializing. So better to
stick to json or whatever at app level.